+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/main/components/investigation_section.tsx
@@ -33,6 +33,7 @@ import { FLYOUT_ORIGIN } from '../../../../common/lib/telemetry';
 import { INVESTIGATION_SECTION_TITLE } from '../../../shared/constants/flyout_titles';
 import { isRulePreviewDocument } from '../../../shared/utils/is_rule_preview_document';
 import { getClusterQualifiedIndex } from '../../../shared/utils/get_cluster_qualified_index';
+import { LEGACY_ANCESTOR_INDEX } from '../constants/field_names';
 
 export const INVESTIGATION_SECTION_TEST_ID = `${PREFIX}InvestigationSection` as const;
 
@@ -84,7 +85,7 @@ export const InvestigationSection = memo(
     const ancestorsIndexName = useMemo(
       () =>
         getClusterQualifiedIndex(
-          (getFieldValue(hit, 'signal.ancestors.index') as string) ?? '',
+          (getFieldValue(hit, LEGACY_ANCESTOR_INDEX) as string) ?? '',
           documentIndexName
         ),
       [hit, documentIndexName]
