+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/tools/session_view/hooks/use_session_view_config.ts
@@ -7,6 +7,7 @@
 
 import { useMemo } from 'react';
 import { type DataTableRecord, getFieldValue } from '@kbn/discover-utils';
+import { getClusterQualifiedIndex } from '../../../../shared/utils/get_cluster_qualified_index';
 import { ALERT_ORIGINAL_TIME } from '../../../../../../common/field_maps/field_names';
 import {
   ANCESTOR_INDEX,
@@ -20,7 +21,11 @@ import type { SessionViewConfig } from '../../../../../../common/types/session_v
  * Hook that returns the session view configuration if the session view is available for the alert
  */
 export const useSessionViewConfig = (hit: DataTableRecord): SessionViewConfig | null => {
-  const index = (getFieldValue(hit, ANCESTOR_INDEX) as string) ?? hit.raw._index;
+  const documentIndexName = hit.raw._index ?? (getFieldValue(hit, '_index') as string) ?? '';
+  const index = getClusterQualifiedIndex(
+    (getFieldValue(hit, ANCESTOR_INDEX) as string) ?? hit.raw._index,
+    documentIndexName
+  );
   const entryLeaderEntityId = getFieldValue(hit, ENTRY_LEADER_ENTITY_ID) as string;
   const entryLeaderStart = getFieldValue(hit, ENTRY_LEADER_START) as string;
   const entityId = getFieldValue(hit, PROCESS_ENTITY_ID) as string;
